refactor, add import alias for types folder
