+++ frontend/src/App.tsx
@@ -4,7 +4,7 @@ export default function App() {
     return (
         <main>
             <header> Header </header>
-            <div className="mx-auto max-w-2xl">
+            <div className="mx-auto max-w-3xl">
                 <Outlet />
             </div>
             <footer> Footer </footer>
