+++ frontend/src/components/room/RoomDetailsCard.tsx
