+++ frontend/src/components/room/RoomEqupmentList.tsx
