+++ frontend/src/services/axios.ts
@@ -1,9 +1,9 @@
-import axios from 'axios';
+import axios from "axios";
 
 // Axios config for api
 export const api = axios.create({
     baseURL: import.meta.env.VITE_API_URL,
     headers: {
-        'Content-Type': 'application/json',
+        "Content-Type": "application/json",
     },
 });
